+++ src/main/kotlin/org/kalasim/examples/elevator/Elevator.kt
@@ -1,4 +1,6 @@
 //Elevator.kt
+@file:OptIn(NumericDuration::class)
+
 package org.kalasim.examples.elevator
 
 import org.kalasim.*
@@ -6,6 +8,7 @@ import org.kalasim.animation.AnimationComponent
 import org.kalasim.examples.elevator.Car.DoorState.CLOSED
 import org.kalasim.examples.elevator.Car.DoorState.OPEN
 import org.kalasim.examples.elevator.Direction.*
+import org.kalasim.misc.NumericDuration
 import org.kalasim.misc.repeat
 import java.awt.Point